refactor: extract pending-edit apply logic into usePendingEditApplicator hook (closes #476)
Move applyCodeChange useCallback (~290 LOC) plus useConfirmDialog,
useToast, and related store selectors out of EditorPanel into a
dedicated usePendingEditApplicator hook.

EditorPanel.tsx: 1003 → 708 lines (−295)
usePendingEditApplicator.ts: 249 lines (new)
